utils and appsettings
