--- TODO.txt
@@ -1,10 +1,0 @@
-done...watch echarts option
-done...sticky ball
-current time pointer
-修正返回值不是json的接口
-import iScroll
-appsettings in a file
-general method code in standalone file
-weather table components realtime module, 5 day module
-use local storage to speed up startup->localstore->netrequest->localstore
-auto location, default location setting
