+++ pages/index/index.vue
@@ -128,10 +128,9 @@
 </template>
 
 <script>
-	import {
-		mapState,
-		mapMutations
-	} from 'vuex'
+	import { mapState, mapMutations } from 'vuex'
+	import appsettings from '../../utils/appsettings.js'
+	import utils from '../../utils/utils.js'
 	import * as echarts from 'echarts'
 	import mpvueEcharts from 'mpvue-echarts'
 
@@ -176,7 +175,7 @@
 					sndballLeft: false, // 第二个球是否位于左端
 					trdballActive: false, // 第三个球是否激活（显示为蓝色）
 					trdballMove: false, // 第三个球是否滑动
-					trdballLeft: false, // 第三个球是否位于左端
+					trdballLeft: false // 第三个球是否位于左端
 				},
 				// 一浴日期球控制参数
 				ballYY: {
@@ -186,7 +185,7 @@
 					sndballLeft: false,
 					trdballActive: false,
 					trdballMove: false,
-					trdballLeft: false,
+					trdballLeft: false
 				},
 				fivedayWeather: [],
 				fivedayHighTemp: [],
@@ -209,7 +208,7 @@
 			loadWeather() {
 				let that = this
 				uni.request({
-					url: 'http://123.234.129.238:8001/MyWebService.asmx/GetCityWeather',
+					url: appsettings.hosturl + 'GetCityWeather',
 					data: {
 						weather: that.array[that.location],
 						name: 'admin',
@@ -231,9 +230,12 @@
 						// 天气情况
 						that.weatherData.weather = result.result.data.realtime.weather.info
 						// 天气图标
-						that.weatherData.weatherIcon = that.setWeatherIcon(that.weatherData.weather)
+						that.weatherData.weatherIcon = utils.setWeatherIcon(
+							that.weatherData.weather
+						)
 						// 空气质量图标及pm2.5字体颜色
-						that.setAirconIcon(that.weatherData.airconDesc)
+						that.weatherData.airconIcon = utils.setAirconIcon(that.weatherData.airconDesc)
+						that.weatherData.pm25Style = utils.setAirconClass(that.weatherData.airconDesc)
 						// 五日天气预报数组，高低温数组
 						let fivedayarr = result.result.data.weather
 						// 清空数组
@@ -248,7 +250,7 @@
 							// 天气 取day的天气
 							let weather = fivedayarr[i].info.day[1]
 							// 天气图标
-							let weatherIcon = that.setWeatherIcon(weather)
+							let weatherIcon = utils.setWeatherIcon(weather)
 							// 风向
 							let windDir = fivedayarr[i].info.day[3]
 							// 风力
@@ -273,95 +275,13 @@
 					} // end-success-request
 				}) // end-request
 			},
-			// 根据天气设置图标
-			setWeatherIcon(weather) {
-				switch (weather) {
-					case '晴':
-					case '晴转多云':
-						return '../../static/Images/right_weather_fine.png'
-						break
-					case '多云':
-					case '多云转晴':
-					case '多云转阵雨':
-					case '多云转阴':
-					case '多云转小雨-中雨':
-						return '../../static/Images/right_weather_cloudy.png'
-						break
-					case '阴':
-					case '阴转小雨-中雨':
-						return '../../static/Images/right_weather_sky.png'
-						break
-					case '雾':
-					case '雾转多云':
-						return '../../static/Images/right_weather_fog.png'
-						break
-					case '风':
-						return '../../static/Images/right_weather_wind.png'
-						break
-					case '小雨':
-					case '中雨':
-					case '中雨-大雨':
-					case '小到中雨':
-					case '小雨转多云':
-					case '小雨-中雨转多云':
-					case '小雨-中雨转小雨':
-					case '大雨':
-					case '暴雨':
-					case '中到暴雨':
-					case '小到暴雨':
-						return '../../static/Images/right_weather_rainy.png'
-						break
-					case '阵雨':
-					case '阵雨转多云':
-						return '../../static/Images/right_weather_rainysh.png'
-						break
-					case '大到暴雨':
-						return '../../static/Images/right_weather_rainstorm.png'
-						break
-					case '雷雨':
-					case '雷阵雨':
-						return '../../static/Images/right_weather_thunder.png'
-						break
-					case '雪':
-						return '../../static/Images/right_weather_snow.png'
-						break
-					case '雨加雪':
-						return '../../static/Images/right_weather_raisnow.png'
-						break
-				}
-			},
-			// 根据空气质量设置图标和文字style
-			setAirconIcon(airconDesc) {
-				switch (airconDesc) {
-					case '重度污染':
-						this.weatherData.airconIcon =
-							'../../static/Images/right_leaf_sev.png'
-						this.weatherData.pm25Style = 'sev'
-						break
-					case '轻度污染':
-						this.weatherData.airconIcon =
-							'../../static/Images/right_leaf_bad.png'
-						this.weatherData.pm25Style = 'bad'
-						break
-					case '良':
-						this.weatherData.airconIcon =
-							'../../static/Images/right_leaf_very.png'
-						this.weatherData.pm25Style = 'very'
-						break
-					case '优':
-						this.weatherData.airconIcon =
-							'../../static/Images/right_leaf_excellent.png'
-						this.weatherData.pm25Style = 'excellent'
-						break
-				} // end-switch (airconDesc)
-			},
 			// 读取服务器台风和海浪警报
 			loadWarning() {
 				let that = this
 				// 读取台风信息
 				// 请求服务器台风列表
 				uni.request({
-					url: 'http://123.234.129.238:8001/MyWebService.asmx/GetTyphoonList',
+					url: appsettings.hosturl + 'GetTyphoonList',
 					data: {
 						areaflg: '青岛',
 						Typhoonyear: new Date().getFullYear()
@@ -377,7 +297,7 @@
 						let lasttyphoon = typhoonlist[typhoonlist.length - 1].split(',')
 						// 请求last typhoon详情数据
 						uni.request({
-							url: 'http://123.234.129.238:8001/MyWebService.asmx/GetTyphoonPath',
+							url: appsettings.hosturl + 'GetTyphoonPath',
 							data: {
 								areaflg: '青岛',
 								typhoonNumber: lasttyphoon[0] // 台风编号
@@ -410,7 +330,7 @@
 				}) // end-request 请求服务器台风列表
 				// 请求服务器海浪预警列表
 				uni.request({
-					url: 'http://123.234.129.238:8001/MyWebService.asmx/GetEachWarning_ceshi',
+					url: appsettings.hosturl + 'GetEachWarning_ceshi',
 					data: {
 						name: 'admin',
 						areaflg: '青岛'
@@ -443,7 +363,7 @@
 			loadAstronomicalTide() {
 				let that = this
 				uni.request({
-					url: 'http://123.234.129.238:8001/MyWebService.asmx/GetAstronomicalTide_QD',
+					url: appsettings.hosturl + 'GetAstronomicalTide_QD',
 					data: {
 						name: 'admin',
 						areaflg: '山东'
@@ -464,291 +384,27 @@
 							}
 						} // end-for
 						// 由数组生成echarts所需的option
-						that.optionJST = that.setChartOption(arrJST)
-						that.optionYY = that.setChartOption(arrYY)
-					}, // end-success-request
+						that.optionJST = utils.setChartOption(arrJST)
+						that.optionYY = utils.setChartOption(arrYY)
+					} // end-success-request
 				}) // end-request
-			},
-			// 由三日数据生成chart option
-			setChartOption(arr) {
-				// 按日期排序
-				function SortByDate(x, y) {
-					let datex = new Date(x.PREDICTIONDATE)
-					let datey = new Date(y.PREDICTIONDATE)
-					return datex - datey
-				}
-				arr.sort(SortByDate)
-				// 提取潮汐数值
-				let tidedata = [] // 曲线用的数值集
-				let markdata = [] // 最高最低潮竖直标线用的数值集
-				let max = 0 // 三天最高潮位数值
-				let min = 0 // 三天最低潮位数值
-				for (let i = 0; i < arr.length; i++) {
-					// 一次循环为一天的数据
-					// 初始日期时间
-					let inittime = new Date(arr[i].PREDICTIONDATE)
-					// 高低潮需要的日期字符串，必须在这里取值，后面的运算会改变inittime值
-					let datestring = inittime.getFullYear() + '/' + (inittime.getMonth() + 1) + '/' + inittime.getDate()
-					// 每小时数值 加入曲线数值表
-					// 格式为[ [时间1, 数值1], [时间2, 数值2], ... ]
-					tidedata.push([inittime.setHours(inittime.getHours()), arr[i].H0])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H1])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H2])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H3])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H4])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H5])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H6])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H7])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H8])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H9])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H10])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H11])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H12])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H13])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H14])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H15])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H16])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H17])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H18])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H19])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H20])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H21])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H22])
-					tidedata.push([inittime.setHours(inittime.getHours() + 1), arr[i].H23])
-					// 高低潮数值 加入曲线数值表和垂直标线数值表
-					// 第一次高潮
-					if (arr[i].FSTHIGHWIDETIME !== '-') {
-						// 第一次高潮的日期时间
-						let dateFH = new Date(datestring + ' ' + arr[i].FSTHIGHWIDETIME + ':00')
-						// 第一次高潮曲线数据
-						tidedata.push([dateFH.setHours(dateFH.getHours()), arr[i].FSTHIGHWIDEHEIGHT])
-						// 第一次高潮垂直标线数据
-						// 垂直标线数据格式为: [ [{coord: [横坐标1, 纵坐标起始值1]},{coord: [横坐标1, 纵坐标结束值1]}], [{coord: [横坐标2, 纵坐标起始值2]},{coord: [横坐标2, 纵坐标结束值2]}], ... ]
-						let fh = [{
-							coord: [dateFH.setHours(dateFH.getHours()), arr[i].FSTHIGHWIDEHEIGHT]
-						}, {
-							coord: [dateFH.setHours(dateFH.getHours()), 0]
-						}]
-						markdata.push(fh)
-					}
-					// 第一次低潮
-					if (arr[i].FSTLOWWIDETIME !== '-') {
-						// 第一次低潮的日期时间
-						let dateFL = new Date(datestring + ' ' + arr[i].FSTLOWWIDETIME + ':00')
-						// 第一次低潮的曲线数据
-						tidedata.push([dateFL.setHours(dateFL.getHours()), arr[i].FSTLOWWIDEHEIGHT])
-						// 第一次低潮的垂直标线数据
-						let fl = [{
-							coord: [dateFL.setHours(dateFL.getHours()), arr[i].FSTLOWWIDEHEIGHT]
-						}, {
-							coord: [dateFL.setHours(dateFL.getHours()), 0]
-						}]
-						markdata.push(fl)
-					}
-					// 第二次高潮
-					if (arr[i].SCDHIGHWIDETIME !== '-') {
-						// 第二次高潮的日期时间
-						let dateSH = new Date(datestring + ' ' + arr[i].SCDHIGHWIDETIME + ':00')
-						// 第二次高潮的曲线数据
-						tidedata.push([dateSH.setHours(dateSH.getHours()), arr[i].SCDHIGHWIDEHEIGHT])
-						// 第二次高潮的垂直标线数据
-						let sh = [{
-							coord: [dateSH.setHours(dateSH.getHours()), arr[i].SCDHIGHWIDEHEIGHT]
-						}, {
-							coord: [dateSH.setHours(dateSH.getHours()), 0]
-						}]
-						markdata.push(sh)
-					}
-					// 第二次低潮
-					if (arr[i].SCDLOWWIDETIME !== '-') {
-						// 第二次低潮的日期时间
-						let dateSL = new Date(datestring + ' ' + arr[i].SCDLOWWIDETIME + ':00')
-						// 第二次低潮的曲线数据
-						tidedata.push([dateSL.setHours(dateSL.getHours()), arr[i].SCDLOWWIDEHEIGHT])
-						// 第二次低潮的垂直标线数据
-						let sl = [{
-							coord: [dateSL.setHours(dateSL.getHours()), arr[i].SCDLOWWIDEHEIGHT]
-						}, {
-							coord: [dateSL.setHours(dateSL.getHours()), 0]
-						}]
-						markdata.push(sl)
-					}
-					// 计算今天的最高潮位 并与已有的max值比较 将最大值存入max用于水平标线
-					let firstH = Number(arr[i].FSTHIGHWIDEHEIGHT)
-					let secondH = Number(arr[i].SCDHIGHWIDEHEIGHT)
-					if (firstH > secondH) { // 如果First比较大
-						if (firstH > max) {
-							max = firstH
-						}
-					} else { // 如果Second比较大
-						if (secondH > max) {
-							max = secondH
-						}
-					}
-					// 计算今天的最低潮位 并于已有的min值比较 将最小值存入min用于水平标线
-					let firstL = Number(arr[i].FSTLOWWIDEHEIGHT)
-					let secondL = Number(arr[i].SCDLOWWIDEHEIGHT)
-					if (firstL < secondL) { // 如果First比较小
-						if (firstL < min || min === 0) {
-							min = firstL
-						}
-					} else { // 如果Second比较小
-						if (secondL < min || min === 0) {
-							min = secondL
-						}
-					}
-				} // end-for
-				// 排序曲线数据集
-				function SortByFirst(x, y) {
-					return x[0] - y[0]
-				}
-				tidedata.sort(SortByFirst)
-
-				// 数据准备完毕，生成并返回echarts用option
-				let option = {
-					// 图表距离外围div的padding
-					grid: {
-						top: '4%',
-						left: '0%',
-						right: '2%',
-						bottom: '20%',
-						containLabel: true
-					},
-					// 横坐标轴
-					xAxis: {
-						type: 'time',
-						axisLabel: { // 横坐标刻度数值
-							show: true,
-							inside: true,
-							rotate: 90,
-							formatter: function (value, index) {
-								// 格式化为'月-日'，只在第一个刻度显示年份
-								let date = new Date(value)
-								let texts = [(date.getMonth() + 1), date.getDate()];
-								if (index === 0) {
-									texts.unshift(date.getFullYear())
-								}
-								return '\n' + texts.join('-')
-							} // end-formatter-axisLabel
-						} // end-axisLabel
-					},
-					yAxis: {
-						show: false,
-						boundaryGap: ['20%', '20%']	// 纵坐标轴的范围，比有效数字上下多出20%
-					},
-					series: [
-						// 第一组series： 曲线数据 + 高低潮垂直标线 + 标线顶部数字label
-						{
-							name: '潮汐',
-							type: 'line',
-							smooth: true,
-							silent: true,
-							animation: false,
-							symbolSize: 0,			// 曲线上数据点小圆圈的大小
-							lineStyle: {
-								color: '#1c8d3b',	// 曲线颜色
-								width: 1			// 曲线粗细
-							},
-							data: tidedata,
-							markLine: {
-								symbolSize: 0.1,	// 垂直标线一端的箭头 和数据label的大小， 不能设为0否则label不显示
-								silent: true,
-								animation: false,
-								label: {
-									show: true,
-									position: 'start',
-									formatter: function (param) {
-										return param.data.coord[1] + 'cm'
-									},
-									/*
-									textStyle: {
-										color: '#000000',
-									}
-									*/
-								}, // end-label-markLine
-								lineStyle: {
-									type: 'dot'
-								},
-								data: markdata
-							}
-						},
-						// 第二组series: 高低潮垂直标线（透明度为0） + 标线底部时间label
-						{
-							name: '标线时间',
-							type: 'line',
-							markLine: {
-								symbolSize: 0.1,
-								opacity: 0,	// 透明度为0 不渲染这条标线
-								silent: true,
-								animation: false,
-								label: {
-									show: true,
-									position: 'end',
-									formatter: function (param) {
-										// 返回mm:ss格式的时间
-										let date = new Date(param.data.coord[0])
-										let hour = date.getHours()
-										if (hour < 10) {
-											hour = '0' + hour
-										}
-										let minute = date.getMinutes()
-										if (minute < 10) {
-											minute = '0' + minute
-										}
-										return hour + ':' + minute
-									},
-									/*
-									textStyle: {
-										color: '#000000',
-									}
-									*/
-								}, // end-label-markLine
-								lineStyle: {
-									type: 'dot'
-								},
-								data: markdata
-							} // end-markLine
-						},
-						// 第三组series： 两条水平标线 表示三天最高和最低的潮位
-						{
-							name: '最值横线',
-							type: 'line',
-							markLine: {
-								symbolSize: 0,
-								silent: true,
-								animation: false,
-								lineStyle: {
-									type: 'dot'
-								},
-								label: {
-									show: false
-								},
-								data: [
-									{ yAxis: max },
-									{ yAxis: min },
-								]
-							} // end-markLine
-						}
-					] // end-series
-				} // end-option
-				return option
 			},
 			// 初始化金沙滩图表
 			handleInitJST(canvas, width, height) {
-				chartJST = echarts.init(canvas, null, {
+				(chartJST = echarts.init(canvas, null, {
 					width: width,
 					height: height
-				}),
+				})),
 					canvas.setChart(chartJST)
 				chartJST.setOption(this.optionJST)
 				return chartJST
 			},
 			// 初始化一浴图表
 			handleInitYY(canvas, width, height) {
-				chartYY = echarts.init(canvas, null, {
+				(chartYY = echarts.init(canvas, null, {
 					width: width,
 					height: height
-				}),
+				})),
 					canvas.setChart(chartYY)
 				chartYY.setOption(this.optionYY)
 				return chartYY
@@ -764,7 +420,8 @@
 			// 设置日期球的状态 scrollLeft为滚动距最左边的距离，ballObj为包含一系列bool值的object
 			setDateballStatus(scrollLeft, ballObj) {
 				//开始滚动 scrollLeft为0
-				if (scrollLeft < 50) { // 刚开始滚动 还不足以让第二个球开始动
+				if (scrollLeft < 50) {
+					// 刚开始滚动 还不足以让第二个球开始动
 					ballObj.fstballActive = true
 					ballObj.sndballActive = false
 					ballObj.trdballActive = false
@@ -772,7 +429,8 @@
 					ballObj.sndballLeft = false
 					ballObj.trdballMove = false
 					ballObj.trdballLeft = false
-				} else if (scrollLeft < this.systemInfo.windowWidth - 45) { // 第二个球开始动
+				} else if (scrollLeft < this.systemInfo.windowWidth - 45) {
+					// 第二个球开始动
 					ballObj.fstballActive = true
 					ballObj.sndballActive = false
 					ballObj.trdballActive = false
@@ -780,7 +438,8 @@
 					ballObj.sndballLeft = false
 					ballObj.trdballMove = false
 					ballObj.trdballLeft = false
-				} else if (scrollLeft < this.systemInfo.windowWidth) { // 第二个球停在最左边 第三个球还没开始动
+				} else if (scrollLeft < this.systemInfo.windowWidth) {
+					// 第二个球停在最左边 第三个球还没开始动
 					ballObj.fstballActive = false
 					ballObj.sndballActive = true
 					ballObj.trdballActive = false
@@ -788,7 +447,8 @@
 					ballObj.sndballLeft = true
 					ballObj.trdballMove = false
 					ballObj.trdballLeft = false
-				} else if (scrollLeft < this.systemInfo.windowWidth + 9) { // 第三个球开始动
+				} else if (scrollLeft < this.systemInfo.windowWidth + 9) {
+					// 第三个球开始动
 					ballObj.fstballActive = false
 					ballObj.sndballActive = true
 					ballObj.trdballActive = false
@@ -796,7 +456,8 @@
 					ballObj.sndballLeft = true
 					ballObj.trdballMove = true
 					ballObj.trdballLeft = false
-				} else if (scrollLeft < (this.systemInfo.windowWidth + 270)) { // 第三个球动
+				} else if (scrollLeft < this.systemInfo.windowWidth + 270) {
+					// 第三个球动
 					ballObj.fstballActive = false
 					ballObj.sndballActive = true
 					ballObj.trdballActive = false
@@ -804,7 +465,8 @@
 					ballObj.sndballLeft = true
 					ballObj.trdballMove = true
 					ballObj.trdballLeft = false
-				} else { // 第三个球停在最左边
+				} else {
+					// 第三个球停在最左边
 					ballObj.fstballActive = false
 					ballObj.sndballActive = false
 					ballObj.trdballActive = true
@@ -817,7 +479,8 @@
 			// 设置曲线图下方日期球的日期
 			setDateballText() {
 				let now = new Date()
-				function formatDate(date) { // 格式化日期为MM-dd
+				function formatDate(date) {
+					// 格式化日期为MM-dd
 					let month = date.getMonth() + 1
 					let day = date.getDate()
 					if (month < 10) {
@@ -834,7 +497,7 @@
 				this.sndballText = formatDate(now)
 				now = new Date(now.setDate(now.getDate() + 1))
 				this.trdballText = formatDate(now)
-			},
+			}
 		}, // end-methods
 		watch: {
 			// 金沙滩option
