+++ utils/appsettings.js
@@ -1,0 +1,5 @@
+const hosturl = 'http://123.234.129.238:8001/MyWebService.asmx/'
+
+module.exports = {
+    hosturl: hosturl
+}
